Shader: Define a common interface for running vertex shader programs.
